+++ hc/api/tests/test_notify_signal.py
@@ -95,7 +95,9 @@ class NotifySignalTestCase(BaseTestCase):
         self.assertEqual(n.error, "")
 
         params = socketobj.req["params"]
-        self.assertIn("“Daily Backup” is DOWN", params["message"])
+        self.assertIn("Daily Backup is DOWN", params["message"])
+        self.assertEqual(params["textStyle"][0], "10:12:BOLD")
+
         self.assertIn("Project: Alices Project", params["message"])
         self.assertIn("Tags: foo, bar", params["message"])
         self.assertIn("Period: 1 day", params["message"])
@@ -138,7 +140,7 @@ class NotifySignalTestCase(BaseTestCase):
         self.assertEqual(n.error, "")
 
         params = socketobj.req["params"]
-        self.assertIn("“Foo & Co” is DOWN", params["message"])
+        self.assertIn("Foo & Co is DOWN", params["message"])
         self.assertIn("Project: Alice & Friends", params["message"])
         self.assertIn("Tags: foo, a&b", params["message"])
 
@@ -328,6 +330,9 @@ class NotifySignalTestCase(BaseTestCase):
         }
         setup_mock(socket, msg)
 
+        self.check.name = "Foo & Co"
+        self.check.save()
+
         self.channel.notify(self.check)
 
         n = Notification.objects.get()
@@ -343,8 +348,15 @@ class NotifySignalTestCase(BaseTestCase):
         email = emails["alice@example.org"]
         self.assertEqual(
             email.subject,
-            "Signal notification failed: The check “Daily Backup” is DOWN.",
+            "Signal notification failed: The check Foo & Co is DOWN.",
         )
+        # The plaintext version should have no HTML markup, and should
+        # have no &amp;, &lt; &gt; stuff:
+        self.assertIn("The check Foo & Co is DOWN.", email.body)
+        # The HTML version should retain styling, and escape special characters
+        # in project name, check name, etc.:
+        html = email.alternatives[0][0]
+        self.assertIn("The check <b>Foo &amp; Co</b> is <b>DOWN</b>.", html)
 
     @patch("hc.api.transports.socket.socket")
     def test_it_handles_null_data(self, socket):