+++ hc/lib/html.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import re
+from html import unescape
 from html.parser import HTMLParser
 
 
@@ -34,3 +36,42 @@ def html2text(html, skip_pre=False):
 
     parser.feed(html)
     return parser.get_text()
+
+
+def extract_signal_styles(markup: str) -> tuple[str, list[str]]:
+    """Convert HTML syntax to Signal text styles.
+
+    This implementation has limited functionality, and only supports the features
+    we do use:
+    * only supports <b> and <code> tags
+    * does not support nested (<b><code>text</code></b>) tags
+
+    Example:
+
+    >>> extract_signal_styles("<b>foo</b> bar")
+    "foo bar", ["0:3:BOLD"]
+
+    """
+    text = ""
+    styles: list[str] = []
+    tag, tag_idx = "", 0
+
+    for part in re.split(r"(</?(?:b|code)>)", markup):
+        if part == "<b>":
+            tag = "BOLD"
+            tag_idx = len(text)
+        elif part == "</b>":
+            assert tag == "BOLD"
+            len_tagged = len(text) - tag_idx
+            styles.append(f"{tag_idx}:{len_tagged}:{tag}")
+        elif part == "<code>":
+            tag = "MONOSPACE"
+            tag_idx = len(text)
+        elif part == "</code>":
+            assert tag == "MONOSPACE"
+            len_tagged = len(text) - tag_idx
+            styles.append(f"{tag_idx}:{len_tagged}:{tag}")
+        else:
+            text += unescape(part)
+
+    return text, styles